# gin-backend/docs/swagger.yaml
basePath: /
definitions:
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  entities.Ciudadano:
    properties:
      alias:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
    type: object
  entities.CiudadanoListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.Ciudadano'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.CiudadanoMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.CiudadanoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Ciudadano'
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.CreateCiudadanoRequest:
    properties:
      alias:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - alias
    - email
    - password
    type: object
...
  entities.LoginCiudadanoRequest:
    properties:
      email_or_alias:
        type: string
      password:
        type: string
    required:
    - email_or_alias
    - password
    type: object
  entities.LoginCiudadanoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Ciudadano'
      message:
        type: string
      success:
        type: boolean
      token:
        type: string
    type: object
  entities.UpdateCiudadanoRequest:
    properties:
      alias:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/ciudadanos:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CiudadanoListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar ciudadanos
      tags:
      - Ciudadano
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateCiudadanoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.CiudadanoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Crear ciudadano
      tags:
      - Ciudadano
  /api/ciudadanos/{id}:
    delete:
      parameters:
      - description: ID del ciudadano
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CiudadanoMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar ciudadano
      tags:
      - Ciudadano
    get:
      parameters:
      - description: ID del ciudadano
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CiudadanoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener ciudadano por ID
      tags:
      - Ciudadano
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID del ciudadano
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateCiudadanoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CiudadanoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar ciudadano
      tags:
      - Ciudadano
  /api/ciudadanos/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciales
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.LoginCiudadanoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.LoginCiudadanoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Login de ciudadano
      tags:
      - Ciudadano
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token JWT con el prefijo Bearer. Ejemplo: Bearer eyJhbGciOiJIUzI1NiIs...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"